Bump the minor version for release, updating the group too
Renaming the entire source tree now that this is under reiseburo
