--- src/test/groovy/com/github/lookout/verspaetung/KafkaPollerSpec.groovy
@@ -1,6 +1,0 @@
-package com.github.lookout.verspaetung
-
-import spock.lang.*
-
-class KafkaPollerSpec extends Specification {
-}
