--- src/test/groovy/com/github/lookout/verspaetung/zk/AbstractTreeWatcherSpec.groovy
@@ -1,25 +1,0 @@
-package com.github.lookout.verspaetung.zk
-
-import spock.lang.*
-
-import com.github.lookout.verspaetung.TopicPartition
-import org.apache.curator.framework.CuratorFramework
-import org.apache.curator.framework.recipes.cache.ChildData
-import org.apache.curator.framework.recipes.cache.TreeCacheEvent
-import org.apache.curator.framework.recipes.cache.TreeCacheEvent
-
-class AbstractTreeWatcherSpec extends Specification {
-    private AbstractTreeWatcher watcher
-
-    class MockWatcher extends AbstractTreeWatcher {
-        MockWatcher() {
-            super(null, [:])
-        }
-        void childEvent(CuratorFramework c, TreeCacheEvent e) { }
-        String zookeeperPath() { return '/zk/spock' }
-    }
-
-    def setup() {
-        this.watcher = new MockWatcher()
-    }
-}
