+++ src/test/groovy/com/github/reiseburo/verspaetung/KafkaPollerSpec.groovy
@@ -1,0 +1,6 @@
+package com.github.reiseburo.verspaetung
+
+import spock.lang.*
+
+class KafkaPollerSpec extends Specification {
+}
